Improve PlayerController and NavGrid
Fix capitalization of nodeDrawHeightOffsetAmount.
Add a Raycast method for finding the position where a ray intersects the NavGrid.
Add a GetClosestWalkableNode method.
Remove Linq from GetClosestNode.
Change PlayerController to use GetClosestWalkableNode for finding the closest node to the clicked location.
Change PlayerController to only find a new path if the destination is not null.
Rename PlayerController.graph to grid.